pyproject.toml
[project]
name = "dlt-test"
version = "0.1.0"
description = "DLT test project"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alive-progress>=3.3.0",
    "apache-airflow[pandas,postgres]>=3.2.1",
    "click>=8.3.3",
    "connectorx>=0.4.5",
    "dbt>=1.0.0.40.17",
    "dlt[cli,duckdb,postgres,sql-database]>=1.26.0",
    "duckdb>=1.5.2",
    "psycopg2-binary>=2.9.12",
    "pyarrow>=24.0.0",
    "pytest>=9.0.3",
    "pytest-lazy-fixtures>=1.4.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = ["-s", "-v", "-p", "no:warnings"]

[dependency-groups]
dev = []
